feat: Enhance @explain command with diagram capability
Fixes #1879
